feat: refactor combat.py
Resolves #71

Factored into combat_hooks and combat
